upgraded script to support reverse tunnels
what does this mean? reverse runneling is a way to
allow us to ssh to a device by ssh'ing to another
device. crazy, huh! the way it works is as follows

port not accessible      port accessible
[firewalled device] <--- [server with public ip]
                    ^ this is the reverse tunnel

in this scenario, the IP we are connecting to via
ssh is NOT necessarily the IP of the device we want
to Mosh into. hence we must go through a public IP
discovery using SSH and ensure to bind the mosh
server to the correct IP address in this case.